refactor(ipod): extract backlight logic to hook
